Resolve integer overflow issue
Resolve integer overflow issue when doing fft in aac decoder. We need 
to add some transform when doing add operation between two LONG values.

Reviewed-on: https://gerrit.mot.com/1296543
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Submit-Approved: Jira Key

+++ libFDK/src/fft.cpp
@@ -106,14 +106,16 @@ amm-info@iis.fraunhofer.de
 #define W_PiFOURTH STC(0x5a82799a)
 //#define W_PiFOURTH ((FIXP_DBL)(0x5a82799a))
 #ifndef SUMDIFF_PIFOURTH
+/*BEGIN, Motorola zhangtai3, IKSWP-40593*/
 #define SUMDIFF_PIFOURTH(diff, sum, a, b) \
   {                                       \
     FIXP_DBL wa, wb;                      \
     wa = fMultDiv2(a, W_PiFOURTH);        \
     wb = fMultDiv2(b, W_PiFOURTH);        \
     diff = wb - wa;                       \
-    sum = wb + wa;                        \
+    sum = (LONG)(((INT64)wb) + wa);       \
   }
+/*END, IKSWP-40593*/
 #define SUMDIFF_PIFOURTH16(diff, sum, a, b)       \
   {                                               \
     FIXP_SGL wa, wb;                              \
